+++ NEWS.md
@@ -5,9 +5,10 @@ TO DO:
 * revamp plotting methods (e.g. use autoplot and so on)
 * vignettes need to be updated to reflect all changes from the refactoring
 * code coverage, eventually, once all of the above is done
-* anything that comes up in the meanwhile
 * check docs
 * rebuild pkgdown website
+* anything that comes up in the meanwhile
+* how method to return call to create plots
 --> targeting rsimsum 0.5.0 to go on CRAN ASAP
 
 * Updated vignettes (NEED TO DO THE PLOTTING ONE STILL) after refactoring.